server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
gradio-client==0.7.0
langdetect==1.0.9
groq==0.4.1
httpx==0.25.0
pydantic==2.5.0